refactor: split database schema by domain (#55)
Organize schema into domain modules for better maintainability:
- auth.ts: BetterAuth tables (user, session, account, verification)
- storage.ts: Nexus tables (storageTierEnum, fileStatusEnum, files)
- index.ts: re-exports all tables/enums

Cross-domain references import directly from source (storage imports user from auth).
Drizzle config updated to use glob pattern for automatic discovery.
